back_python/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.109.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.14
starlette==0.35.1
typing_extensions==4.15.0
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
